requirements.txt
django
djangorestframework
drf-yasg
python-dotenv
psycopg2-binary
openai
pgvector
llama-index
llama-parse
nest_asyncio
djangorestframework-api-key
django-cors-headers
